+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/configs/BzClientConfigs.java
@@ -24,4 +24,5 @@ public class BzClientConfigs {
     public static float essenceItemHUDVisualEffectSpeed = 1;
     public static boolean showBeesOnGuiAllYearRound = false;
     public static boolean showBeesOnGuiOnAprilFools = true;
+    public static boolean maximumBeesOnGui = false;
 }