+++ fabric-base/src/main/java/com/telepathicgrunt/the_bumblezone/configs/fabricbase/BzConfig.java
@@ -445,6 +445,9 @@ public class BzConfig extends MidnightConfig {
     @Entry
     public static boolean showBeesOnGuiOnAprilFools = true;
 
+    @Entry
+    public static boolean maximumBeesOnGui = false;
+
 
     @ApiStatus.Internal
     public static void setup() {
@@ -595,5 +598,6 @@ public class BzConfig extends MidnightConfig {
         BzClientConfigs.essenceItemHUDVisualEffectSpeed = essenceItemHUDVisualEffectSpeed;
         BzClientConfigs.showBeesOnGuiAllYearRound = showBeesOnGuiAllYearRound;
         BzClientConfigs.showBeesOnGuiOnAprilFools = showBeesOnGuiOnAprilFools;
+        BzClientConfigs.maximumBeesOnGui = maximumBeesOnGui;
     }
 }