+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### **(V.7.10.2 Changes) (1.20.1 Minecraft)*
+### **(V.7.11.0 Changes) (1.20.1 Minecraft)*
 
 #### Entities:
 Added a Bee Queen trade to exchange Banner Patterns for Experience Bottles.
@@ -8,6 +8,7 @@ Added a Bee Queen trade to exchange Banner Patterns for Experience Bottles.
 Added an April Fool's joke for future years, where bees can appear on any in-world gui.
  This is controlled by three configs: showBeesOnGuiAllYearRound, showBeesOnGuiOnAprilFools, and maximumBeesOnGui
  You can activate these gui bees 24/7 with the year round config!
+ The Bumblezone Teleporting screen will have these gui bees always as a small thing if stuck waiting for dimension to load.
 
 
 ### **(V.7.10.1 Changes) (1.20.1 Minecraft)*